fix(bigint): RFC-0110 bug fixes - Knuth division, borrow/carry, canonicalization
Critical fixes:
- limb_sub: Fix borrow propagation using overflowing_sub chain
- limb_mul: Fix carry propagation (replace |= with proper loop)
- bigint_divmod: Replace buggy binary search with Knuth Algorithm D

High severity fixes:
- bigint_shr: Return ZERO for large shifts (not Err)
- bigint_shl: Handle shift == 0 correctly
- Serialization: Fix byte offsets (num_limbs at byte 4)
- Add is_canonical() checks to all arithmetic functions

Lower severity:
- Add BigIntError::OutOfRange variant
- Add debug_assert for non-empty limbs in BigInt::new()
- Narrow #[allow(arithmetic_overflow)] scope in probe.rs

Regression tests: 69 tests added, 2 previously slow division tests
now enabled (Knuth Algorithm D is O(n) not O(quotient)).

+++ determin/src/probe.rs
@@ -1,5 +1,4 @@
 #![allow(dead_code)]
-#![allow(arithmetic_overflow)]
 
 //! Deterministic Floating-Point Verification Probe
 //!
@@ -528,7 +527,7 @@ pub fn bigint_all_probe_entries() -> Vec<BigIntProbeEntry> {
         BigIntProbeEntry {
             index: 1,
             op_id: OP_ADD,
-            input_a: BigIntProbeValue::Int(1 << 64),
+            input_a: BigIntProbeValue::Int((1u128 << 64) as i128),
             input_b: BigIntProbeValue::Int(1),
             description: "2^64 + 1",
         },
